requirements.txt
tencentcloud-sdk-python==3.0.302
requests==2.22.0